+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 0.2 (unreleased)
 ----------------
 
+- Allow explicit configuration of all cluster ports.
+
 - Added an `address` property to the cluster, to expose a client connection
   string including all cluster nodes.